updated i'm bored interacitons
changed the dialog box for 'I'm bored' on the newsfeed page a bit to
have more input boxes for the user. Also started some preliminary
actions for when the user submits the same form. The newsfeed now
updates with a new activity. We will still need to send the info in the
form to the server and let the server add the new activity and send json
back to the client to populate the new activity item on success. made
some minor changes to the css to make some things look better or
function properly with the changes. Also added normalize.css to help
with cross browser consistency.